Move RequestState base classes to be inner classes of Resolver.
