--- reasoner/resolution/framework/RequestState.java
@@ -1,79 +1,0 @@
-/*
- * Copyright (C) 2021 Vaticle
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package com.vaticle.typedb.core.reasoner.resolution.framework;
-
-import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
-import com.vaticle.typedb.core.common.poller.Poller;
-import com.vaticle.typedb.core.concept.answer.ConceptMap;
-import com.vaticle.typedb.core.reasoner.resolution.answer.AnswerState;
-import com.vaticle.typedb.core.reasoner.resolution.framework.Resolver.DownstreamManager;
-
-import java.util.HashSet;
-import java.util.Optional;
-import java.util.Set;
-
-import static com.vaticle.typedb.core.common.poller.Pollers.poll;
-
-public abstract class RequestState {
-
-    protected final Request.Visit fromUpstream;
-
-    protected RequestState(Request.Visit fromUpstream) {
-        this.fromUpstream = fromUpstream;
-    }
-
-    public abstract Optional<? extends AnswerState.Partial<?>> nextAnswer();
-
-    public interface Exploration {
-
-        boolean newAnswer(AnswerState.Partial<?> partial);
-
-        DownstreamManager downstreamManager();
-
-        boolean singleAnswerRequired();
-
-    }
-
-    public abstract static class CachingRequestState<ANSWER> extends RequestState {
-
-        protected final AnswerCache<ANSWER> answerCache;
-        protected Poller<? extends AnswerState.Partial<?>> cacheReader;
-        protected final Set<ConceptMap> deduplicationSet;
-
-        protected CachingRequestState(Request.Visit fromUpstream, AnswerCache<ANSWER> answerCache, boolean deduplicate) {
-            super(fromUpstream);
-            this.answerCache = answerCache;
-            this.deduplicationSet = deduplicate ? new HashSet<>() : null;
-            this.cacheReader = answerCache.reader().flatMap(
-                    a -> poll(toUpstream(a).filter(
-                            partial -> !deduplicate || !deduplicationSet.contains(partial.conceptMap()))));
-        }
-
-        @Override
-        public Optional<? extends AnswerState.Partial<?>> nextAnswer() {
-            Optional<? extends AnswerState.Partial<?>> ans = cacheReader.poll();
-            if (ans.isPresent() && deduplicationSet != null) deduplicationSet.add(ans.get().conceptMap());
-            return ans;
-        }
-
-        protected abstract FunctionalIterator<? extends AnswerState.Partial<?>> toUpstream(ANSWER answer);
-
-    }
-
-}
